refactor: Separate helpers from interfaces
