+++ internal/core/interfaces.go
@@ -31,13 +31,3 @@ func NewIntegration(config config.Integration) (Integration, error) {
 	}
 	return integ, nil
 }
-
-func IsSource(i Integration) (Source, bool) {
-	s, ok := i.(Source)
-	return s, ok
-}
-
-func IsDestination(i Integration) (Destination, bool) {
-	d, ok := i.(Destination)
-	return d, ok
-}
